# feat(backend): optimize text chunking performance and large document handling (#132)